Game changes
1) gitattributes - file consistency
2) Pinky Implementation
3) Fixed orange and red (int) errors
4) Purple Implementation - DLS
5) Game - screen scaling + sound
6) Changed A* Default in Level file
7) VoidTile/vector fixed for collisions
8) Sound file plays now

+++ pacman/src/game/entity/mob/OrangeGhost.java
@@ -91,8 +91,8 @@ public class OrangeGhost extends Mob {
 		if (path != null && path.size() > 0) {
 			// A* returns path from goal back to start, so last element is the next step
 			Vector2i next = path.get(path.size() - 1).tile;
-			int nextPx = next.getX() << 4;
-			int nextPy = next.getY() << 4;
+			int nextPx = (int) next.getX() << 4;
+			int nextPy = (int) next.getY() << 4;
 
 			if (x < nextPx)
 				xa++;
